\name{calc_seconds}
\alias{calc_seconds}
\title{Calculate seconds from character strings in specified formats}
\usage{
  calc_seconds(x)
}
\arguments{
  \item{x}{Character vector of times in the MedAssociates
  format (e.g., 0000:00.00)}
}
\description{
  Calculate seconds from character strings in specified
  formats
}

