\name{list_zones}
\alias{list_zones}
\title{List zone names contained in zone file}
\usage{
  list_zones(txt, zone.define = "Zones Defined")
}
\arguments{
  \item{zone.define}{header label denoting zone definition
  area}

  \item{txt}{character vector containing each line of the
  raw .zone file}
}
\description{
  List zone names contained in zone file
}

