\name{parse_meta}
\alias{parse_meta}
\title{Parse metadata contained between subject header and zone definition section}
\usage{
  parse_meta(txt)
}
\arguments{
  \item{txt}{character vector containing each line of the
  raw .zone file}
}
\value{
  A data frame containing each subject's associated meta
  data.
}
\description{
  Parse metadata contained between subject header and zone
  definition section
}

