\name{parse_zonefile}
\alias{parse_zonefile}
\title{Read in and parse zone file}
\usage{
  parse_zonefile(file, convert.times = TRUE)
}
\arguments{
  \item{file}{Zone file}

  \item{convert.times}{use \code{\link{convert_time}} to
  automatically identify MedAssociates time formats and
  convert values to numeric seconds}
}
\description{
  Read in and parse zone file
}

