\name{split_cols}
\alias{split_cols}
\title{Split character vectors and return data frame}
\usage{
  split_cols(txt, split)
}
\arguments{
  \item{split}{regular expression pattern used to delimit
  columns}

  \item{txt}{character vector containing each line of the
  raw .zone file}
}
\value{
  Data frame
}
\description{
  Split character vectors and return data frame
}

