\name{calc_seconds}
\alias{calc_seconds}
\title{Calculate seconds from MedAssociates time formats}
\usage{
  calc_seconds(x)
}
\arguments{
  \item{x}{Character vector of times in the MedAssociates
  format (e.g., 0000:00.00)}
}
\value{
  numeric vector of time in seconds.
}
\description{
  Calculate seconds from MedAssociates time formats
}

